refactor: :art: better naming convention
rename to include hooks for easier findability
