+++ frontend/src/helpers/agreement-hooks.js
