+++ frontend/src/pages/agreements/details/AgreementBudgetLines.js
@@ -5,7 +5,7 @@ import AgreementDetailHeader from "./AgreementDetailHeader";
 import { CreateBudgetLinesProvider } from "../../../components/UI/WizardSteps/StepCreateBudgetLines/context";
 import BudgetLinesTable from "../../../components/UI/BudgetLinesTable";
 import StepCreateBudgetLines from "../../../components/UI/WizardSteps/StepCreateBudgetLines/StepCreateBudgetLines";
-import { useIsUserAllowedToEditAgreement } from "../../../helpers/useAgreements";
+import { useIsUserAllowedToEditAgreement } from "../../../helpers/agreement-hooks";
 
 /**
  * Renders Agreement budget lines view
