+++ frontend/src/pages/agreements/details/AgreementDetails.js
@@ -4,7 +4,7 @@ import AgreementValuesCard from "../../../components/Agreements/AgreementDetails
 import AgreementDetailHeader from "./AgreementDetailHeader";
 import AgreementDetailsView from "./AgreementDetailsView";
 import AgreementDetailsEdit from "./AgreementDetailsEdit";
-import { useIsAgreementEditable } from "../../../helpers/useAgreements";
+import { useIsAgreementEditable } from "../../../helpers/agreement-hooks";
 
 /**
  * Renders the details of an agreement, including budget lines, spending, and other information.
